recipe.md
Soften butter. If you are planning on making these, take the butter out of the fridge first thing in the morning so it’s ready to go when you need it.

Measure the flour correctly. Be sure to use a measuring cup made for dry ingredients (NOT a pyrex liquid measuring cup). There has been some controversy on how to measure flour. I personally use the scoop and shake method and always have (gasp)! It’s easier and I have never had that method fail me.  Many of you say that the only way to measure flour is to scoop it into the measuring cup and level with a knife. I say, measure it the way you always do. Just make sure that the dough matches the consistency of the dough in the photos in this post.

Use LOTS of chocolate chips

DO NOT over-bake these chocolate chip cookies! I explain this more below, but these chocolate chip cookies will not look done when you pull them out of the oven, and that is GOOD.
When you remove the cookies from the oven they will still look doughy. THIS is the secret that makes these cookies so absolutely amazing! Please, I beg you, do NOT overbake!
**Butter. I use Kirkland Brand Salted Butter from Costco to make these cookies. I have also used Tillamook salted butter with equally excellent results. Unsalted butter would also be great. I just recommend tasting the dough to ensure that it's salted to your liking.
***Some people have said they think the cookies are too salty. Please be sure to use natural Sea Salt (not iodized table salt). If you are concerned about saltiness start with 1/2 tsp salt and adjust to your tastes. I always make it with salted butter and 1 tsp sea salt. 
**Recipe information calculated based on this recipe making 36 cookies - 2 TBS of dough apiece)
All-purpose flour: Many readers have used gluten-free all-purpose flour with excellent results.